// src/models/IFilter.ts
export enum Category {
    SNEACKERS = 'sneacker',
    WEAR = 'wear',
}

export enum Brand {
    ADIDAS = 'Adidas',
    YEEZY = 'Yeezy',
    NIKE = 'Nike',
    AIR_JORDAN = 'Air Jordan',
    NEW_BALANCE = 'New Balance',
}

export enum Color {
    BLACK = 'Черный',
    WHITE = 'Белый',
    MULTI = 'Многоцветный',
    PINK = 'Розовый',
}

export enum Price {
    K10 = 'Ниже 10 000 РУБ',
    K15 = '10 000 - 15 000 РУБ',
    K20 = '15 000 - 20 000 РУБ',
    K25 = '20 000 - 25000 РУБ',
    K25more = '25000+',
}

export enum Filters {
    BRAND = 'brand',
    COLOR = 'color',
    PRICE = 'price',
}